Rewrite App into a function component in preparation for hooks.
